kauai: pin SwapBytesRglw call sites + Template _pggcmid to int32_t
SwapBytesRglw is misleadingly named — its body is hard-coded 4-byte word
arithmetic (`pb += 4`, `pb[3]` swap, etc) and the runtime
`Assert(size(long) == 4, "code broken")` makes the contract explicit.
But many callers expressed their count as `cb / size(long)`, which is
correct on x86 (size(long)==4) but under-counts by half on LP64 (size==8),
silently swapping only half of every byte-order conversion target.

Swept the codebase:
- Documented the 4-byte-word semantics on the function declaration.
- Replaced runtime `Assert(size(long) == 4)` with `static_assert` against
  size(int32_t) so the contract is checked at compile time.
- Replaced the 11 caller divisor sites computing `cb / size(long)` with
  `cb / size(int32_t)`. Sites: chcm.cpp:991, chunk.cpp (3 sites),
  kidworld.cpp:73, rtxt.cpp:1197, actrsave.cpp:390, bkgd.cpp (3 sites),
  tmpl.cpp:179/195, movie_chomp.cpp (2 sites).
- AssertBomRglw similarly converted from size(long) divisor to size(int32_t).
- Template _pggcmid: GG of (4-byte count + 4-byte cmid array). Pinned the
  PggNew(size(long)) creator to PggNew(size(int32_t)) and the matching
  CbFixed assertion + cast sites — its on-disk form is 4-byte cmids
  regardless of architecture.

Build clean on x86, 33/33 geometry tests pass, chomp + movie-chomp end-to-end
exercise the chunky-file write path through the patched code.

+++ kauai/src/chcm.cpp
@@ -988,7 +988,7 @@ void Compiler::_ParseBodyCursor(bool fPack, ChunkTagOrType ctg, ChunkNumber cno)
 
         // The bits are stored in upside down DIB order!
         ReversePb(pcurh + 1, pcurdir->cb - size(CURH));
-        SwapBytesRglw(pcurh + 1, (pcurdir->cb - size(CURH)) / size(long));
+        SwapBytesRglw(pcurh + 1, (pcurdir->cb - size(CURH)) / size(int32_t));
 
         if (pcurdir->dxp == 16)
         {
